+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Support passing max header and body sizes to client #74
 - Allow additional links to be included #60
 - Resolve relative URLs to absolute URLs #64
+- Show referrer title and Xpath #49
 
 ## [0.5.0] 2019-02-23